Link navigation inside sidebar to render contents
Add new directory data where each data to be displayed of each navigation
element of the sidebar are handled. Each time the data is published, by
event listeners, the content is subscribed to the event so its display
will be updated accordingly.
Only the Inbox and the projects are linked yet.