Fix button section render
The confirm and cancel buttons are moved to "ui/buttons" were they may
be used by more stuff in the future.
Change the idea to display the confirmation for the deletion of a task
inside the same container as the task. Instead, the idea would be to
display an alert of a custom popup to ask confirmation before deletion.